+++ src/private/package/GetNovaPackageFileName.ps1
@@ -2,18 +2,24 @@ function Get-NovaPackageFileName {
     [CmdletBinding()]
     param(
         [Parameter(Mandatory)][pscustomobject]$ProjectInfo,
-        [Parameter(Mandatory)][string]$PackageId
+        [Parameter(Mandatory)][string]$PackageId,
+        [Parameter(Mandatory)][string]$PackageType
     )
 
+    $packageType = ConvertTo-NovaPackageType -Type $PackageType
     $packageFileName = "$( $ProjectInfo.Package.PackageFileName )".Trim()
     if ( [string]::IsNullOrWhiteSpace($packageFileName)) {
-        $packageFileName = "$PackageId.$( $ProjectInfo.Version ).nupkg"
+        $packageFileName = "$PackageId.$( $ProjectInfo.Version )"
     }
 
-    if (-not $packageFileName.EndsWith('.nupkg', [System.StringComparison]::OrdinalIgnoreCase)) {
-        $packageFileName = "$packageFileName.nupkg"
+    $packageFileName = $packageFileName -replace '(?i)(?:\.nupkg|\.zip)$', ''
+    $fileExtension = if ($packageType -eq 'Zip') {
+        '.zip'
+    }
+    else {
+        '.nupkg'
     }
 
-    return $packageFileName
+    return "$packageFileName$fileExtension"
 }